refactor(runtime): restore source size boundaries
