+++ scripts/ci/source-file-size-baseline.json
@@ -29,7 +29,7 @@
     "src/features/runtime-provider-management/main/infrastructure/cli-companion/RuntimeProviderCliCompanionService.ts": 838,
     "src/features/runtime-provider-management/main/infrastructure/OpenCodeLocalProviderConnector.ts": 1018,
     "src/features/runtime-provider-management/renderer/hooks/useRuntimeProviderManagement.ts": 2118,
-    "src/features/runtime-provider-management/renderer/RuntimeLocalProviderSetupDialog.tsx": 1678,
+    "src/features/runtime-provider-management/renderer/RuntimeLocalProviderSetupDialog.tsx": 1676,
     "src/features/runtime-provider-management/renderer/RuntimeProviderQuickConnect.tsx": 826,
     "src/features/runtime-provider-management/renderer/ui/RuntimeProviderManagementPanelView.tsx": 3309,
     "src/features/terminal-workspace/renderer/ui/TerminalWorkspacePanel.tsx": 4513,
@@ -77,7 +77,7 @@
     "src/main/services/team/provisioning/TeamProvisioningStreamEvents.ts": 1212,
     "src/main/services/team/ReviewApplierService.ts": 2062,
     "src/main/services/team/ReviewDecisionStore.ts": 1970,
-    "src/main/services/team/runtime/OpenCodeTeamRuntimeAdapter.ts": 1650,
+    "src/main/services/team/runtime/OpenCodeTeamRuntimeAdapter.ts": 1648,
     "src/main/services/team/TaskChangeComputer.ts": 1357,
     "src/main/services/team/TaskChangeLedgerReader.ts": 2159,
     "src/main/services/team/taskLogs/stream/BoardTaskLogStreamService.ts": 2516,
@@ -96,7 +96,7 @@
     "src/main/services/team/TeamTaskReader.ts": 882,
     "src/main/services/team/TeamTranscriptProjectResolver.ts": 1558,
     "src/main/utils/atomicWrite.ts": 822,
-    "src/main/utils/childProcess.ts": 1047,
+    "src/main/utils/childProcess.ts": 965,
     "src/main/utils/jsonl.ts": 957,
     "src/main/workers/team-fs-worker.ts": 2151,
     "src/preload/constants/ipcChannels.ts": 843,
