+++ src/features/runtime-provider-management/renderer/RuntimeLocalProviderSetupDialog.tsx
@@ -49,6 +49,7 @@ import { LocalProviderBrandIcon } from './ui/LocalProviderBrandIcon';
 import { LocalProviderModelAssignmentControls } from './ui/LocalProviderModelAssignmentControls';
 import { LocalProviderPrivateNetworkApprovalControl } from './ui/LocalProviderPrivateNetworkApprovalControl';
 import { LocalProviderScopeSelector } from './ui/LocalProviderScopeSelector';
+import { RuntimeLocalProviderInlineError as InlineError } from './ui/RuntimeLocalProviderInlineError';
 import { RuntimeProviderEndpointCredentialsFields } from './ui/RuntimeProviderEndpointCredentialsFields';
 import { RuntimeProviderModelTestResult } from './ui/RuntimeProviderModelTestResult';
 import {
@@ -205,16 +206,6 @@ const SetupProgress = ({
     </ol>
   );
 };
-
-const InlineError = ({ message }: { readonly message: string }): JSX.Element => (
-  <div
-    role="alert"
-    className="flex items-start gap-2 rounded-r-md border-l-2 border-red-400/70 bg-red-400/[0.06] px-3 py-2.5 text-xs text-red-200"
-  >
-    <AlertTriangle className="mt-0.5 size-3.5 shrink-0" aria-hidden="true" />
-    <span>{message}</span>
-  </div>
-);
 
 interface RuntimeLocalProviderSetupDialogProps {
   readonly open: boolean;
